Friday Commit
This commit comes after editing every error I could on w3c Validator last night and this morning. Yesterday my github link wouldn't work. The only errors that are showing up now on w3c validator are the bad elements for when I am describing to the user how to, for example, "	<h2> Where To Put Javascript </h2>
		<p> Simple Script- <br /> -Executed when the page loads, place in < script > < / script> tags within the body.</p> "

the < script > is showing up as an error, but Im unsure how else to write it in.

in this commit, my images are probably not going to load either, so I am in the process of figuring out their paths and why they will not work. It's probably something to do with my file management skills..